Версия 1.2.0 Составные атрибуты и некоторые другие вещи поправлена библиотека libfloader
git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@552 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocclient/kksfactory/KKSIndFactory.cpp
@@ -1356,7 +1356,10 @@ void KKSIndFactory :: setValue (QWidget *aw, const KKSIndicator * pCategAttr, co
 
 void KKSIndFactory :: connectToSlots (QObject *aw, QWidget* wEditor)
 {
-    QObject::connect (aw, SIGNAL (valueChanged(int, KKSIndAttr::KKSIndAttrClass, QVariant)), wEditor, SLOT (setIndValue (int, KKSIndAttr::KKSIndAttrClass, QVariant)) );
+    QObject::connect (aw, 
+                      SIGNAL (valueChanged(int64, KKSIndAttr::KKSIndAttrClass, QVariant)), 
+                      wEditor, 
+                      SLOT (setIndValue (qint64, KKSIndAttr::KKSIndAttrClass, QVariant)) );
 }
 
 void KKSIndFactory :: viewIOIndicator (KKSObject * io, int idIndicator)
