+++ ddocclient/kksgui/KKSComboBox.h
@@ -28,7 +28,7 @@ class _GUI_EXPORT KKSComboBox : public QComboBox, public KKSAttrWidget
         //void setItemData(int index);
 
     signals:
-        void valueChanged (int id, KKSIndAttr::KKSIndAttrClass sys, QVariant val);
+        void valueChanged (qint64 id, KKSIndAttr::KKSIndAttrClass sys, QVariant val);
 
     protected:
