+++ ddocclient/kksgui/KKSPointTable.h
@@ -42,7 +42,7 @@ private slots:
     void delCurrPoint (void);
 
 signals:
-    void valueChanged (int id, KKSIndAttr::KKSIndAttrClass isSys, QVariant val);
+    void valueChanged (qint64 id, KKSIndAttr::KKSIndAttrClass isSys, QVariant val);
 
 protected:
     //
