+++ ddocclient/kksgui/KKSText.h
@@ -24,7 +24,7 @@ class _GUI_EXPORT KKSText : public QTextEdit, public KKSAttrWidget
         virtual ~KKSText (void);
 
     signals:
-        void valueChanged (int id, KKSIndAttr::KKSIndAttrClass isSys, QVariant val);
+        void valueChanged (qint64 id, KKSIndAttr::KKSIndAttrClass isSys, QVariant val);
 
     public slots:
